+++ sources/webcore/src/canopsis-next/src/components/other/alarm/actions/actions-panel.vue
@@ -145,10 +145,22 @@ export default {
         afterSubmit: () => this.fetchAlarmsListWithPreviousParams({ widgetId: this.widget._id }),
       };
     },
-    actions() {
+    isResolvedAlarm() {
+      return [ENTITIES_STATUSES.off, ENTITIES_STATUSES.cancelled].includes(this.item.v.status.val);
+    },
+    resolvedActions() {
+      const actions = [];
+
+      if (this.widget.parameters.moreInfoTemplate !== '') {
+        actions.push(this.filteredActionsMap.moreInfos);
+      }
+
+      return actions;
+    },
+    unresolvedActions() {
       const { filteredActionsMap } = this;
 
-      let actions = [
+      const actions = [
         filteredActionsMap.snooze,
         filteredActionsMap.pbehaviorAdd,
         filteredActionsMap.pbehaviorList,
@@ -183,17 +195,23 @@ export default {
         }
       }
 
+      return actions;
+    },
+    actions() {
+      let actions = this.isResolvedAlarm ? this.resolvedActions : this.unresolvedActions;
+
       actions = compact(actions);
 
-      const inlineActions = actions.slice(0, 3);
-      const dropDownActions = actions.slice(3);
-
       const result = {
-        inline: inlineActions.filter(action => !!action),
-        dropDown: dropDownActions.filter(action => !!action),
+        inline: actions.slice(0, 3),
+        dropDown: actions.slice(3),
       };
 
-      if (featuresService.has('components.alarmListActionPanel.computed.actions')) {
+      /**
+       * If we will have actions for resolved alarms in the features we should move this condition to
+       * the every features repositories
+       */
+      if (!this.isResolvedAlarm && featuresService.has('components.alarmListActionPanel.computed.actions')) {
         return featuresService.call('components.alarmListActionPanel.computed.actions', this, result);
       }