+++ sources/webcore/src/canopsis-next/src/mixins/entities/pbehavior-reasons.js
@@ -1,0 +1,15 @@
+import { createNamespacedHelpers } from 'vuex';
+
+const { mapActions, mapGetters } = createNamespacedHelpers('pbehaviorReasons');
+
+export default {
+  computed: {
+    ...mapGetters({
+      pbehaviorReasonsPending: 'pending',
+      pbehaviorReasons: 'pbehaviorReasons',
+    }),
+  },
+  methods: {
+    ...mapActions(['fetchPbehaviorReasons']),
+  },
+};