+++ sources/webcore/src/canopsis-next/src/router.js
@@ -20,6 +20,7 @@ import ExploitationWebhooks from '@/views/exploitation/webhooks.vue';
 import ExploitationSnmpRules from '@/views/exploitation/snmp-rules.vue';
 import ExploitationActions from '@/views/exploitation/actions.vue';
 import ExploitationHeartbeats from '@/views/exploitation/heartbeats.vue';
+import ExploitationDynamicInfos from '@/views/exploitation/dynamic-infos.vue';
 
 Vue.use(Router);
 
@@ -161,6 +162,17 @@ const routes = [
       requiresLogin: true,
       requiresRight: {
         id: USERS_RIGHTS.technical.exploitation.heartbeat,
+      },
+    },
+  },
+  {
+    path: '/exploitation/dynamic-infos',
+    name: 'exploitation-dynamic-infos',
+    component: ExploitationDynamicInfos,
+    meta: {
+      requiresLogin: true,
+      requiresRight: {
+        id: USERS_RIGHTS.technical.exploitation.dynamicInfo,
       },
     },
   },